jpaRemapping values (#5)
* Kunde: dispo und guthaben von BigInteger zu long

* BI2Long: Gewinnklasse:betrag, Lottoscheinziehung:gewinnsuper6+spiel77

* BI2Long Lottoschein: belegnummer

* BI2Long Lottoscheinziehung6aus49: gewinn

* BI2Long Ziehung: zahlenalsbits, einsatzspiel77+super6+lotto

+++ src/main/java/de/wbstraining/lotto/business/lottogesellschaft/AuftragKunde.java
@@ -1,7 +1,6 @@
 package de.wbstraining.lotto.business.lottogesellschaft;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -38,11 +37,11 @@ public class AuftragKunde implements Serializable {
 		return kunde.getVorname();
 	}
 
-	public BigInteger getGuthaben() {
+	public long getGuthaben() {
 		return kunde.getGuthaben();
 	}
 
-	public BigInteger getDispo() {
+	public long getDispo() {
 		return kunde.getDispo();
 	}
